Reverted to an older version w/same template
The copied template implementation caused various issues
with our shady dom implementation; this uses polymer itself more
and relies less on the polymer implementation and thus should be
more-reliable in core polymer updates. It is also less-complex and
has a lot-fewer edge cases. Steve gave a thumbs up to reusing the
template as a preferred method.