Changed to simpler algorithm implementation (#343)
* Reverted to an older version w/same template

The copied template implementation caused various issues
with our shady dom implementation; this uses polymer itself more
and relies less on the polymer implementation and thus should be
more-reliable in core polymer updates. It is also less-complex and
has a lot-fewer edge cases. Steve gave a thumbs up to reusing the
template as a preferred method.

* added proper tests for projection

* copies inert template modifies that using old algo

* forgot to build

* IE compatibility, simpler, replacement finder

* added tests to check that bindings still work

* proper annotation tests