Polymer Flush beforeEach added to a11ySuite
Without Polymer.dom.flush, fixtures with elements containing dom-repeat or dom-if templates would not be fully loaded. In many cases, this results in running the test suite on only a fragment of the page, missing possible errors. In few cases, it would trigger possible errors, for example:

  <div role='radiogroup'>
    <template is='dom-repeat'>
       <ul role='radio'>Button</ul>
    </template> 
  </div>

The test suite would incorrectly think the radiogroup had no role='radio' children.

Also, this adds in a beforeEach function that can be passed in, allowing any extra setup to run before each test.