Fix styling for html and body, so that unnecessary scrollbars don't appear anymore.
For example user agent styles in Chrome include body: { margin: 8px }.
Thus, html doesn't have enough space for body, and both horizontal and
vertical appear.

Scrollbars are annoying when writing screenshot tests, as if the
component takes all the available width (or/and hight), scrollbars
appear on the screenshots too.